+++ tools/generator/ui/src/main/java/is/codion/tools/generator/ui/DomainGeneratorPanel.java
@@ -543,7 +543,7 @@ public final class DomainGeneratorPanel extends JPanel {
 										.controls(createMainMenuControls())
 										.buildMenuBar())
 						.defaultCloseOperation(WindowConstants.EXIT_ON_CLOSE)
-						.centerFrame(true)
+						.center(true)
 						.show();
 	}